Refactor [form schema] task form
