+++ pages/edu/assignment_schema/edit/[spec].tsx
@@ -1,6 +1,6 @@
 import { useLocale } from '@hooks/useLocale';
-import { ReactNode, useCallback, useEffect, useMemo } from 'react';
-import { useForm, UseFormReturnType } from '@mantine/form';
+import { ReactNode, useCallback, useMemo } from 'react';
+import { UseFormReturnType } from '@mantine/form';
 import { ITaskDisplay } from '@custom-types/data/ITask';
 import { DefaultLayout } from '@layouts/DefaultLayout';
 import { Item } from '@ui/CustomTransferList/CustomTransferList';
