+++ pages/group/edit/[spec].tsx
@@ -1,7 +1,7 @@
 import Form from '@components/Group/Form/Form';
 import { useLocale } from '@hooks/useLocale';
 import { ReactNode, useCallback, useMemo } from 'react';
-import { useForm, UseFormReturnType } from '@mantine/form';
+import { UseFormReturnType } from '@mantine/form';
 import { DefaultLayout } from '@layouts/DefaultLayout';
 import { getApiUrl } from '@utils/getServerUrl';
 import { GetStaticPaths, GetStaticProps } from 'next';
