+++ pages/task/add.tsx
@@ -1,6 +1,6 @@
 import { useLocale } from '@hooks/useLocale';
 import { DefaultLayout } from '@layouts/DefaultLayout';
-import { useForm } from '@mantine/form';
+import { UseFormReturnType } from '@mantine/form';
 import { ReactNode, useCallback, useEffect, useState } from 'react';
 import { useUser } from '@hooks/useUser';
 import { ITaskDisplay } from '@custom-types/data/ITask';
@@ -64,77 +64,6 @@ function AddTask() {
     IHintAlarmType[]
   >([]);
 
-  const form = useForm({
-    initialValues,
-    validate: {
-      title: (value) =>
-        value.length < 5 ? locale.task.form.validation.title : null,
-      tags: (value) =>
-        value.length == 1 ? locale.task.form.validation.tags : null,
-      description: (value) =>
-        value.length < 20
-          ? locale.task.form.validation.description
-          : null,
-      inputFormat: (value) =>
-        value.length == 0
-          ? locale.task.form.validation.inputFormat
-          : null,
-      outputFormat: (value) =>
-        value.length == 0
-          ? locale.task.form.validation.outputFormat
-          : null,
-      constraintsMemory: (value) =>
-        value < 0 || value > 1024
-          ? locale.task.form.validation.constraints.memory
-          : null,
-      constraintsTime: (value) =>
-        value < 0.5 || value > 30
-          ? locale.task.form.validation.constraints.time
-          : null,
-      complexity: (value) =>
-        value < 0
-          ? locale.task.form.validation.complexity.least
-          : value > 100
-          ? locale.task.form.validation.complexity.most
-          : null,
-      examples: (value, values) =>
-        value.length < 1
-          ? locale.task.form.validation.examples.number
-          : value.filter(
-              (pair: ITest) =>
-                pair.inputData.trim() || pair.outputData.trim()
-            ).length != value.length
-          ? locale.task.form.validation.examples.empty
-          : null,
-      tests: (value, values) =>
-        value.length < 1
-          ? locale.task.form.validation.tests.number
-          : value.filter(
-              (pair: ITest) =>
-                (pair.inputData.trim() && values.taskType == '0') ||
-                (pair.outputData.trim() && values.taskType == '0')
-            ).length != value.length
-          ? locale.task.form.validation.tests.empty
-          : null,
-      checkerCode: (value, values) =>
-        values.checkType == '1' && value.length == 0
-          ? locale.task.form.validation.checkerCode
-          : null,
-      hintContent: (value, values) =>
-        values.hasHint
-          ? value.length == 0
-            ? locale.task.form.validation.hintContent
-            : null
-          : null,
-      hintAlarm: (value, values) =>
-        values.hasHint
-          ? value < 0
-            ? locale.task.form.validation.hintAlarm
-            : null
-          : null,
-    },
-  });
-
   const { data, loading } = useRequest<{}, ITaskAddBundle>(
     'bundle/task_add'
   );
@@ -147,81 +76,84 @@ function AddTask() {
     }
   }, [data]);
 
-  const handleSubmit = useCallback(() => {
-    if (form.validate().hasErrors) {
-      const id = newNotification({});
-      errorNotification({
-        id,
-        title: locale.validationError,
-        autoClose: 5000,
-      });
-      return;
-    }
-    const {
-      checkerCode,
-      checkerLang,
-      hintContent,
-      hintAlarmType,
-      hintAlarm,
-      constraintsMemory,
-      constraintsTime,
-      allowedLanguages,
-      forbiddenLanguages,
-      tags,
-      ...values
-    } = form.values;
-    let body: any = {
-      ...values,
-      author: user?.shortName || 'unknown',
-      checkType: +form.values['checkType'],
-      taskType: +form.values['taskType'],
-      constraints: {
-        time: constraintsTime,
-        memory: constraintsMemory,
-      },
-      allowedLanguages: allowedLanguages.map(
-        (lang: Item) => lang.value
-      ),
-      forbiddenLanguages: forbiddenLanguages.map(
-        (lang: Item) => lang.value
-      ),
-      tags: tags.map((tag: Item) => tag.value),
-    };
-    if (!form.values.shouldRestrictLanguages) {
-      body.allowedLanguages = [];
-      body.forbiddenLanguages = [];
-    }
-    if (form.values['checkType'] === '1') {
-      body.checker = {
-        sourceCode: checkerCode,
-        language: +checkerLang,
+  const handleSubmit = useCallback(
+    (form: UseFormReturnType<any>) => {
+      if (form.validate().hasErrors) {
+        const id = newNotification({});
+        errorNotification({
+          id,
+          title: locale.validationError,
+          autoClose: 5000,
+        });
+        return;
+      }
+      const {
+        checkerCode,
+        checkerLang,
+        hintContent,
+        hintAlarmType,
+        hintAlarm,
+        constraintsMemory,
+        constraintsTime,
+        allowedLanguages,
+        forbiddenLanguages,
+        tags,
+        ...values
+      } = form.values;
+      let body: any = {
+        ...values,
+        author: user?.shortName || 'unknown',
+        checkType: +form.values['checkType'],
+        taskType: +form.values['taskType'],
+        constraints: {
+          time: constraintsTime,
+          memory: constraintsMemory,
+        },
+        allowedLanguages: allowedLanguages.map(
+          (lang: Item) => lang.value
+        ),
+        forbiddenLanguages: forbiddenLanguages.map(
+          (lang: Item) => lang.value
+        ),
+        tags: tags.map((tag: Item) => tag.value),
       };
-    }
-    if (form.values['remark'].trim() === '') {
-      body.remark = undefined;
-    }
-    if (form.values['hasHint']) {
-      body.hint = {
-        content: hintContent,
-        alarmType: +hintAlarmType,
-        alarm: hintAlarm,
-      };
-    }
-    requestWithNotify(
-      'task/add',
-      'POST',
-      locale.notify.task.create,
-      lang,
-      (response: ITaskDisplay) => response.spec,
-      body
-    );
-  }, [form, locale, user?.login, lang]);
+      if (!form.values.shouldRestrictLanguages) {
+        body.allowedLanguages = [];
+        body.forbiddenLanguages = [];
+      }
+      if (form.values['checkType'] === '1') {
+        body.checker = {
+          sourceCode: checkerCode,
+          language: +checkerLang,
+        };
+      }
+      if (form.values['remark'].trim() === '') {
+        body.remark = undefined;
+      }
+      if (form.values['hasHint']) {
+        body.hint = {
+          content: hintContent,
+          alarmType: +hintAlarmType,
+          alarm: hintAlarm,
+        };
+      }
+      requestWithNotify(
+        'task/add',
+        'POST',
+        locale.notify.task.create,
+        lang,
+        (response: ITaskDisplay) => response.spec,
+        body
+      );
+    },
+    [locale, user, lang]
+  );
 
   return (
     <>
       {!loading && (
         <Form
-          form={form}
+          initialValues={initialValues}
           taskTypes={taskTypes}
           taskCheckTypes={taskCheckTypes}
           hintAlarmTypes={hintAlarmTypes}
